Adding functions to get search intervals from a point and a radius
Using an integer based spatial index (geohash) we can get search
intervals from a <lat, lng> and a radius.

We hash the coordinates in a bounding box, then we look for the
neighbors of that box (to get elements right in the border) and return
those.

An optimization is made when two intervals can be merged, so we merge
them to query less times.

GeoHash library: github.com/corsc/go-geohash

Transformation based on:

* http://gis.stackexchange.com/questions/18330/would-it-be-possible-to-use-geohash-for-proximity-searches/92331#92331
* https://github.com/arjunmehta/node-geo-proximity